+++ index.php
@@ -1,196 +1,26 @@
-<!doctype html>
-<html lang="en">
+<!DOCTYPE html>
+<html>
   <head>
-    <!-- Required meta tags -->
     <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <link rel="shortcut icon" href="#" />  
-    <title>Tasti</title>
-      
-    <!-- Bootstrap CSS -->
-    <link rel="stylesheet" href="assets/bootstrap/css/bootstrap.min.css">
-    <!-- CSS personalizado --> 
-    <link rel="stylesheet" href="main.css">  
-      
-      
-    <!--datables CSS básico-->
-    <link rel="stylesheet" type="text/css" href="assets/datatables/datatables.min.css"/>
-    <!--datables estilo bootstrap 4 CSS-->  
-    <link rel="stylesheet"  type="text/css" href="assets/datatables/DataTables-1.10.18/css/dataTables.bootstrap4.min.css">    
-      
-    <link href="https://fonts.googleapis.com/icon?family=Material+Icons"
-      rel="stylesheet">  
+    <title>Welcome to Tasti</title>
+    <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
+    <link rel="stylesheet" href="assets/css/style.css">
   </head>
-    
-  <body> 
-    
 
-    <header>
-    <div class="container">
-        <div class="row">
-            <div class="form-group">    
-                <h3 class='text-center'>TASTI</h3>
-            </div>
-            <div class="form-group">          
-                <button id="btnRegister" type="button" class="btn btn-info" data-toggle="modal">Register</button>      
-            </div>  
-        </div>
-          
-    </div> 
-     
-     </header>    
-    
-    <div class="container">
-        <div class="row">
-            <div class="col-lg-12">            
-            <button id="btnNuevaConsulta" type="button" class="btn btn-info" data-toggle="modal">Nueva Consulta</button>    
-            </div>    
-        </div>    
-    </div>    
-    <br>  
+  <body>
+    <?php require 'template/header.php' ?>
 
-    <div class="container caja">
-        <div class="row">
-            <div class="col-lg-12">
-            <div class="table-responsive">        
-                <table id="tablaConsulta" class="table table-striped table-bordered table-condensed" style="width:100%" >
-                    <thead class="text-center">
-                        <tr>
-                            <th>Id</th>
-                            <th>Usuario</th>
-                            <th>Titulo</th>
-                            <th>Descripcion</th>                                 
-                            <th>Acciones</th>
-                        </tr>
-                    </thead>
-                    <tbody>                           
-                    </tbody>        
-                </table>               
-            </div>
-            </div>
-        </div>  
-    </div>   
+    <?php if(!empty($user)): ?>
+      <br> Welcome. <?= $user['username']; ?>
+      <br>You are Successfully Logged In
+      <a href="template/logout.php">
+        Logout
+      </a>
+    <?php else: ?>
+      <h1>Please Login or SignUp</h1>
 
-<!--Modal para CRUD-->
-<!--<div class="modal fade" id="modalCRUD" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
-    <div class="modal-dialog" role="document">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title" id="exampleModalLabel"></h5>
-                <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span>
-                </button>
-            </div>
-        <form id="formUsuarios">    
-            <div class="modal-body">
-                <div class="row">
-                    <div class="col-lg-6">
-                    <div class="form-group">
-                    <label for="" class="col-form-label">User Name:</label>
-                    <input type="text" class="form-control" id="username">
-                    </div>
-                    </div>
-                    <div class="col-lg-6">
-                    <div class="form-group">
-                    <label for="" class="col-form-label">First Name</label>
-                    <input type="text" class="form-control" id="first_name">
-                    </div> 
-                    </div>    
-                </div>
-                <div class="row"> 
-                    <div class="col-lg-6">
-                    <div class="form-group">
-                    <label for="" class="col-form-label">Last Name</label>
-                    <input type="text" class="form-control" id="last_name">
-                    </div>               
-                    </div>
-                    <div class="col-lg-6">
-                    <div class="form-group">
-                    <label for="" class="col-form-label">Gender</label>
-                    <input type="text" class="form-control" id="gender">
-                    </div>
-                    </div>  
-                </div>
-                <div class="row">
-                    <div class="col-lg-9">
-                        <div class="form-group">
-                        <label for="" class="col-form-label">Password</label>
-                        <input type="text" class="form-control" id="password">
-                        </div>
-                    </div>    
-                    <div class="col-lg-3">    
-                        <div class="form-group">
-                        <label for="" class="col-form-label">Status</label>
-                        <input type="number" class="form-control" id="status">
-                        </div>            
-                    </div>    
-                </div>                
-            </div>
-            <div class="modal-footer">
-                <button type="button" class="btn btn-light" data-dismiss="modal">Cancelar</button>
-                <button type="submit" id="btnGuardar" class="btn btn-dark">Guardar</button>
-            </div>
-        </form>    
-        </div>
-    </div>
-</div>--->
-
-
-
-<div class="modal fade" id="modalCRUD" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
-    <div class="modal-dialog" role="document">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title" id="exampleModalLabel"></h5>
-                <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span>
-                </button>
-            </div>
-        <form id="formConsulta">    
-            <div class="modal-body">
-                <div class="row">
-                    <div class="col-lg-6">
-                        <div class="form-group">
-                        <label for="" class="col-form-label">Usuario:</label>
-                        <input type="text" class="form-control" id="usuario">
-                        </div>
-                    </div>  
-                </div>
-                <div class="row">
-                    <div class="col-lg-6">
-                        <div class="form-group">
-                        <label for="" class="col-form-label">Titulo:</label>
-                        <input type="text" class="form-control" id="titulo">
-                        </div>
-                    </div>  
-                </div>
-                <div class="row"> 
-                    <div class="col-lg-6">
-                        <div class="form-group">
-                        <label for="" class="col-form-label">Descripcion: </label>
-                        <input type="text" class="form-control" id="descripcion">
-                        </div>               
-                    </div>
-                </div>              
-            </div>
-            <div class="modal-footer">
-                <button type="button" class="btn btn-light" data-dismiss="modal">Cancelar</button>
-                <button type="submit" class="btn btn-dark">Guardar</button>
-            </div>
-        </form>    
-        </div>
-    </div>
-</div>
-
-      
-    <!-- jQuery, Popper.js, Bootstrap JS -->
-    <script src="assets/jquery/jquery-3.3.1.min.js"></script>
-    <script src="assets/popper/popper.min.js"></script>
-    <script src="assets/bootstrap/js/bootstrap.min.js"></script>
-      
-    <!-- datatables JS -->
-    <script type="text/javascript" src="assets/datatables/datatables.min.js"></script>    
-     
-    <script type="text/javascript" src="main.js"></script>  
-    
-    
+      <a href="template/login.php">Login</a> or
+      <a href="template/signup.php">SignUp</a>
+    <?php endif; ?>
   </body>
 </html>
