+++ js/validar.js
@@ -1,0 +1,27 @@
+$(document).ready(function() {
+  var user_id, opcion, password;
+  opcion = 10;
+
+  tablaUsuarios = $('#tablaUsuarios').DataTable({  
+    "ajax":{            
+        "url": "bd/crud.php", 
+        "method": 'POST', //usamos el metodo POST
+        "data":{opcion:4}, //enviamos opcion 4 para que haga un SELECT
+        "dataSrc":""
+    },
+    "columns":[
+        {"data": "user_id"},
+        {"data": "username"},
+        {"data": "first_name"},
+        {"data": "last_name"},
+        {"data": "gender"},
+        {"data": "password"},
+        {"data": "status"},
+        {"defaultContent": "<div class='text-center'><div class='btn-group'><button class='btn btn-primary btn-sm btnEditar'><i class='material-icons'>edit</i></button><button class='btn btn-danger btn-sm btnBorrar'><i class='material-icons'>delete</i></button></div></div>"}
+    ]
+  });
+
+  
+
+
+});
