+++ template/login.php
@@ -1,0 +1,35 @@
+<?php
+
+?>
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>Login</title>
+    <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
+    <link rel="stylesheet" href="../assets/css/style.css">
+  </head>
+  <body>
+    <?php require 'header.php' ?>
+
+    <h1>Login</h1>
+    <span>or <a href="signup.php">SignUp</a></span>
+
+    <form action="home.php" method="POST">
+      <input name="email" type="text" placeholder="Enter your username">
+      <input name="password" type="password" placeholder="Enter your Password">
+      <input type="submit" value="Submit">
+    </form>
+  
+    <!-- jQuery, Popper.js, Bootstrap JS -->
+    <script src="../assets/jquery/jquery-3.3.1.min.js"></script>
+    <script src="../assets/popper/popper.min.js"></script>
+    <script src="../assets/bootstrap/js/bootstrap.min.js"></script>
+      
+    <!-- datatables JS -->
+    <script type="text/javascript" src="../assets/datatables/datatables.min.js"></script>    
+     
+    <script type="text/javascript" src="../js/main.js"></script>  
+  
+  </body>
+</html>
