+++ template/logout.php
@@ -1,0 +1,9 @@
+<?php
+  session_start();
+
+  session_unset();
+
+  session_destroy();
+
+  <?php require 'header.php' ?>
+?>
